Moved location utility to `/util/location.ts`
